A lot of changes!
Changed the display and GUI to be run with PYQT instead of OpenCV which will improve quality of life. The number of objects detected per class is now displayed along with the number of detections. Removed asyncio in main thread, capture and display is now still run in the main thread but being done with PYQT without async. Added a fps controller, and settings menu will popup when you press 'K' which allows you to control the fps. Changed the looks of the 'No Feed Detected' text, I think it looks better so its staying. Did some heavy commenting with the `absolute_path` function, it now will search within specified folders within the parent folder.