I don't like AGPL 3.0
I've switched the model to YoloNAS due to the fact that Yolov10 is under a very strict copyleft license, YoloNAS is an overall improvement with better accuracy and inference speeds. I also removed anything to do with ultralytics for the same reasons as Yolov10. Because of my strong urge to not have anything to do with AGPL 3.0 I decided to make my own function that takes care of drawing detection boxes, I made the bboxes switch color and opacity based on confidence scores and created smooth transition to avoid color and opacity flickering.