Added distance estimation
The code has been organized and commented, added distance estimation. Distance estimation uses the objects width and height and compares them with the known width and height of that object to estimate the distance. This result isn't accurate but it's close enough once the camera is calibrated. Distance estimation for a single camera is challenging, more accurate results would come from stereo cameras but these aren't supported in this project so far and most security cameras aren't stereo to begin with.