Feature changes
Removed CUDA for GPU acceleration in favor of DirectML, this allows for more diverse GPU usage. Added the ability to cancel camera calibration, disabled some logs that display in the terminal but the super gradient library had to be modified to remove the sg logs, if running the source code with default libraries, super gradient will still be showing logs in and saving them.