Massive Improvements
The folder structure was completely re-organised for a more easier time with coding, switched to a different and more efficient model, the PP-YOLOE model. Introduced a decode function that decodes the raw outputs from the model and makes the data usable for the drawing functions i made. moved camera capture to a separate thread for faster frame access.